検索フィルター画面リファクタリング
・custom_checkboxをプライベートクラスに変更
・genderをserviceからdataへ移動
・ヘッダーをappbarからnavbarに変更
